LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/drawElementImage-zero-size-expected.txt


FAIL drawElementImage with zero size does nothing assert_array_equals: expected property 0 to be 0 but got 255 (expected array [0, 128, 0, 255] got object "255,0,0,255")
